+++ package.json
@@ -19,6 +19,7 @@
     "commander": "^12.1.0",
     "dayjs": "^1.11.12",
     "dotenv": "^16.4.5",
+    "ink": "^5.0.1",
     "inquirer": "^10.1.2",
     "node-fetch": "^3.3.2",
     "openapi-fetch": "^0.11.1",
